requirements.txt
streamlit==1.31.0
google-generativeai>=0.8.0
pillow>=10.0.0,<11.0.0
huggingface-hub==0.20.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
openai>=1.0.0
nltk>=3.8.0
anthropic>=0.18.0
fpdf2>=2.7.0